Split server-settings-advanced.test.ts (1414→876 lines)
Moves custom domain, event-column-order, and attendee-column-order
tests into server-settings-advanced-custom-domain.test.ts (559 lines)
to bring both files under the 1200-line biome limit.
